+++ ckan/__init__.py
@@ -1,6 +1,6 @@
 # encoding: utf-8
 
-__version__ = '2.9.1'
+__version__ = '2.9.2'
 
 __description__ = 'CKAN Software'
 __long_description__ = \